[fix] refactor presentation layer
