+++ internal/services/event/interface.go
@@ -1,0 +1,11 @@
+package event
+
+import (
+	"context"
+
+	"exchange/internal/domain/notification"
+)
+
+type INotificationService interface {
+	Notify(ctx context.Context, n *notification.Notification) error
+}
